rename firebase client file
