+++ src/lib/client/Feature/Auth/firebase.ts
